requirements.txt
# Core dependencies
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.2.0
chromadb>=0.4.22
sentence-transformers>=2.2.0
ollama>=0.1.8
pyyaml>=6.0

# Document loaders
pypdf>=3.0.0
docx2txt>=0.8
ebooklib>=0.18
unstructured>=0.10.0
lxml>=4.9.0

# Rich terminal UI
rich>=13.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Progress (optional, replaced by rich)
tqdm>=4.66.0
